Implement a test for the Experimentation module
